# web/public/openapi-integration.yaml
...
components:
    parameters:
...
        hookId:
            in: path
            name: hookId
            required: true
            schema:
                type: string
        id:
            in: path
            name: id
            required: true
            schema:
                type: string
...
    responses:
        Error:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
            description: Error response
    schemas:
        AppUser:
            properties:
                attributes:
                    additionalProperties: true
                    type: object
                display_name:
                    type: string
                email:
                    nullable: true
                    type: string
                external_id:
                    nullable: true
                    type: string
                id:
                    type: string
                organisation_id:
                    type: string
                status:
                    enum:
                        - active
                        - disabled
                        - archived
                    type: string
            type: object
...
    securitySchemes:
        bearerAuth:
            bearerFormat: JWT
            description: Session token (kyc_sess_…) or API key / service token
            scheme: bearer
            type: http
info:
    description: |-
        JSON API for merchant backends integrating with KYC (app users, attributes, product plans, entitlements).

        Authenticate with an **organisation API key**: `Authorization: Bearer kyc_…` (create keys under Platform → API keys).
        Keys may be scoped to RBAC permissions; empty scopes grant full organisation access. The organisation must have the `api_access` entitlement.

        Operator UI routes (Google OAuth, members, roles, permissions, email/automation config, platform admin) are documented separately under **Operator API**.
    title: KYC Integration API
    version: 1.0.0
openapi: 3.0.3
paths:
    /v1/app-users/{id}:
        delete:
            operationId: deleteAppUser
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AppUser'
                    description: Deleted app user
                default:
                    $ref: '#/components/responses/Error'
            summary: Delete app user
            tags:
                - App users
        get:
            operationId: getAppUser
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AppUser'
                    description: App user
                default:
                    $ref: '#/components/responses/Error'
            summary: Get app user
            tags:
                - App users
        parameters:
            - $ref: '#/components/parameters/id'
        patch:
            description: When attributes is sent, replaces the entire attributes map.
            operationId: updateAppUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAppUserRequest'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AppUser'
                    description: Updated app user
                default:
                    $ref: '#/components/responses/Error'
            summary: Update app user
            tags:
                - App users
    /v1/attribute-definitions/{id}:
        delete:
            operationId: deleteAttributeDefinition
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AttributeDefinition'
                    description: Archived definition
                default:
                    $ref: '#/components/responses/Error'
            summary: Archive attribute definition
            tags:
                - Attributes
        get:
            operationId: getAttributeDefinition
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AttributeDefinition'
                    description: Definition
                default:
                    $ref: '#/components/responses/Error'
            summary: Get attribute definition
            tags:
                - Attributes
        parameters:
            - $ref: '#/components/parameters/id'
        patch:
            operationId: updateAttributeDefinition
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAttributeDefinitionRequest'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AttributeDefinition'
                    description: Updated definition
                default:
                    $ref: '#/components/responses/Error'
            summary: Update attribute definition
            tags:
                - Attributes
    /v1/entitlements:
        get:
            operationId: listEntitlements
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    items:
                                        items:
                                            $ref: '#/components/schemas/Entitlement'
                                        type: array
                                type: object
                    description: Entitlement list
                default:
                    $ref: '#/components/responses/Error'
            summary: List global entitlement catalog
            tags:
                - Billing
    /v1/entitlements/check:
        post:
            operationId: checkEntitlement
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CheckEntitlementRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    allowed:
                                        type: boolean
                                    reason:
                                        enum:
                                            - disabled
                                            - override_include
                                            - override_exclude
                                            - percentage
                                            - full
                                            - "off"
                                        type: string
                                type: object
                    description: Entitlement check result
                default:
                    $ref: '#/components/responses/Error'
            summary: Check organisation entitlement
            tags:
                - Billing
    /v1/hooks/inbound/{hookId}:
        parameters:
            - $ref: '#/components/parameters/hookId'
        post:
            description: Public endpoint. Auth via X-KYC-Webhook-Secret header or ?secret= / ?token= per auth_mode.
            operationId: receiveInboundWebhook
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/InboundWebhookPayload'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    ok:
                                        type: boolean
                                type: object
                    description: Accepted
                default:
                    $ref: '#/components/responses/Error'
            security: []
            summary: Receive inbound webhook (header or query auth)
            tags:
                - Webhooks
    /v1/hooks/inbound/{hookId}/{token}:
        parameters:
            - $ref: '#/components/parameters/hookId'
            - in: path
              name: token
              required: true
              schema:
                type: string
        post:
            operationId: receiveInboundWebhookWithToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/InboundWebhookPayload'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    ok:
                                        type: boolean
                                type: object
                    description: Accepted
                default:
                    $ref: '#/components/responses/Error'
            security: []
            summary: Receive inbound webhook (path-token auth)
            tags:
                - Webhooks
    /v1/organisations/{id}:
        get:
            operationId: getOrganisation
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Organisation'
                    description: Organisation
                default:
                    $ref: '#/components/responses/Error'
            summary: Get organisation
            tags:
                - Organisations
        parameters:
            - $ref: '#/components/parameters/orgId'
    /v1/organisations/{id}/app-users:
        get:
            operationId: listAppUsers
            parameters:
                - $ref: '#/components/parameters/statusQuery'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    items:
                                        items:
                                            $ref: '#/components/schemas/AppUser'
                                        type: array
                                type: object
                    description: App user list
                default:
                    $ref: '#/components/responses/Error'
            summary: List app users
            tags:
                - App users
        parameters:
            - $ref: '#/components/parameters/orgId'
        post:
            description: Requires app_users:write. attributes keys must match active definitions.
            operationId: createAppUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAppUserRequest'
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AppUser'
                    description: Created app user
                default:
                    $ref: '#/components/responses/Error'
            summary: Create app user
            tags:
                - App users
    /v1/organisations/{id}/app-users/ingest:
        parameters:
            - $ref: '#/components/parameters/orgId'
        put:
            description: |
                Lookup uses org app_user_ingest_upsert_key (external_id or email). attributes are merged.
                discover mode auto-creates unknown attribute definitions; strict rejects unknown keys.
            operationId: ingestAppUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IngestAppUserRequest'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/AppUser'
                                    - properties:
                                        created:
                                            example: false
                                            type: boolean
                                      type: object
                    description: Updated app user
                "201":
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/AppUser'
                                    - properties:
                                        created:
                                            example: true
                                            type: boolean
                                      type: object
                    description: Created app user
                default:
                    $ref: '#/components/responses/Error'
            summary: Upsert app user from external system
            tags:
                - App users
    /v1/organisations/{id}/attribute-definitions:
        get:
            description: Seeds system defaults if missing. Requires attributes:read.
            operationId: listAttributeDefinitions
            parameters:
                - $ref: '#/components/parameters/statusQuery'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    items:
                                        items:
                                            $ref: '#/components/schemas/AttributeDefinition'
                                        type: array
                                type: object
                    description: Definition list
                default:
                    $ref: '#/components/responses/Error'
            summary: List attribute definitions
            tags:
                - Attributes
        parameters:
            - $ref: '#/components/parameters/orgId'
        post:
            description: Requires attributes:manage.
            operationId: createAttributeDefinition
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAttributeDefinitionRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AttributeDefinition'
                    description: Created definition
                default:
                    $ref: '#/components/responses/Error'
            summary: Create attribute definition
            tags:
                - Attributes
    /v1/organisations/{id}/entitlements:
        get:
            operationId: getOrganisationEntitlements
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OrgEntitlements'
                    description: Effective entitlements by scope
                default:
                    $ref: '#/components/responses/Error'
            summary: Get effective organisation entitlements
            tags:
                - Billing
        parameters:
            - $ref: '#/components/parameters/orgId'
    /v1/organisations/{id}/product-features:
        get:
            operationId: listProductFeatures
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    items:
                                        items:
                                            $ref: '#/components/schemas/ProductFeature'
                                        type: array
                                type: object
                    description: Feature list
                default:
                    $ref: '#/components/responses/Error'
            summary: List product features
            tags:
                - Product
        parameters:
            - $ref: '#/components/parameters/orgId'
        post:
            operationId: createProductFeature
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateProductFeatureRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductFeature'
                    description: Created feature
                default:
                    $ref: '#/components/responses/Error'
            summary: Create product feature
            tags:
                - Product
    /v1/organisations/{id}/product-plan:
        get:
            operationId: getActiveProductPlan
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductPlan'
                    description: Active product plan
                default:
                    $ref: '#/components/responses/Error'
            summary: Get active product plan
            tags:
                - Product
        parameters:
            - $ref: '#/components/parameters/orgId'
        put:
            operationId: setActiveProductPlan
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SetActiveProductPlanRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                properties:
                                    product_plan:
                                        allOf:
                                            - $ref: '#/components/schemas/ProductPlan'
                                        nullable: true
                                type: object
                    description: Active plan updated
                default:
                    $ref: '#/components/responses/Error'
            summary: Set active product plan
            tags:
                - Product
...
    /v1/organisations/{id}/subscription:
        get:
            description: Requires billing:read.
            operationId: getSubscription
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Subscription'
                    description: Subscription
                default:
                    $ref: '#/components/responses/Error'
            summary: Get organisation subscription
            tags:
                - Billing
        parameters:
            - $ref: '#/components/parameters/orgId'
...
security:
    - bearerAuth: []
servers:
    - description: Relative to the web app proxy (paths include `/v1`)
      url: /
...